elf: Move dynamice relocation code to separate module
In the effort of splitting up the ELF parser code, move the
dynamic relocation code to its own module.
